feat: release v0.1.0

+++ src/camera_data_store.hpp
@@ -45,7 +45,7 @@ public:
   void reset_camera_images();
 
 private:
-  const int rois_number_;
+  const size_t rois_number_;
   const int image_height_;
   const int image_width_;
   rclcpp::Logger logger_;
